Created skyblockitem endpoint
