+++ src/api/v1/hypixel/skyblockitem.ts
@@ -1,5 +1,4 @@
-import { getBazaarData } from "../../../backend/ItemData";
-import { prisma } from "../../../backend/backend";
+import { getSkyblockItemData } from "../../../backend/itemdata/ItemData";
 import { api } from "../../api";
 
 
@@ -7,7 +6,7 @@ let lastResponse = ""
 let lastReponseTime = 0
 const RESPONSE_CACHE_TIME_MINUTES = 5
 
-export function loadBazaarDataEndpoint() {
+export function loadSkyblockItemEndpoint() {
   api.get('/v1/hypixel/skyblockitem', (req, res) => {
     if (lastReponseTime + RESPONSE_CACHE_TIME_MINUTES * 5 * 1000 > Date.now() ) {
       console.log("Sending Cached Data")
@@ -15,7 +14,7 @@ export function loadBazaarDataEndpoint() {
       res.end()
     } else {
       console.log("Sending New Data")
-      getBazaarData().then( (data) => {
+      getSkyblockItemData().then( (data) => {
         if (data.success != true) {
           res.status(500)
           res.send(data.data)
